+++ accessors/schema_meta.go
@@ -1,0 +1,31 @@
+/*
+Copyright 2015-2021 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package accessors
+
+// SchemaMeta represents schema metadata struct
+type SchemaMeta struct {
+	// Name field name in target struct
+	Name string
+
+	// IsTime field contains time
+	IsTime bool
+
+	// IsDuration field contains duration
+	IsDuration bool
+
+	// Nested nested message definition
+	Nested map[string]*SchemaMeta
+}
